feat: add texts to localization
