+++ src/client/components/target-selection/index.ts
@@ -1,2 +1,3 @@
 export { TargetSelection } from './target-selection';
 export { TargetForSelection } from './target-for-selection';
+export { getTarget } from './get-target';
