+++ src/client/localization/game-screen-localization.ts
@@ -2,4 +2,8 @@ export interface IGameScreenLocalization {
   ReadyButton: string,
   TargetSelection: string,
   ALlEnemies: string,
+  SpellCast: string,
+  RollBonus: string,
+  MakeDiceRoll: string,
+  SelectSpellError: string,
 }
